LibWeb: Split InheritStyleValue out of StyleValue.{h,cpp}
