+++ Userland/Libraries/LibWeb/CSS/StyleValue.h
@@ -627,26 +627,6 @@ private:
     NonnullOwnPtr<CalcSum> m_expression;
 };
 
-class InheritStyleValue final : public StyleValueWithDefaultOperators<InheritStyleValue> {
-public:
-    static ValueComparingNonnullRefPtr<InheritStyleValue> the()
-    {
-        static ValueComparingNonnullRefPtr<InheritStyleValue> instance = adopt_ref(*new InheritStyleValue);
-        return instance;
-    }
-    virtual ~InheritStyleValue() override = default;
-
-    ErrorOr<String> to_string() const override { return "inherit"_string; }
-
-    bool properties_equal(InheritStyleValue const&) const { return true; }
-
-private:
-    InheritStyleValue()
-        : StyleValueWithDefaultOperators(Type::Inherit)
-    {
-    }
-};
-
 class InitialStyleValue final : public StyleValueWithDefaultOperators<InitialStyleValue> {
 public:
     static ValueComparingNonnullRefPtr<InitialStyleValue> the()
